Move reify_dims/reify_arrays to hypercube.util
